Add expressions evaluator

+++ src/big_number/operations/compare.cpp
@@ -1,7 +1,5 @@
 #include "operations.h"
 
-#include <cassert>
-
 namespace big_num::operations {
 
 TSign CompareUnchecked(const TVector &left, const TVector &right) {
@@ -9,7 +7,7 @@ TSign CompareUnchecked(const TVector &left, const TVector &right) {
         return left.size() < right.size() ? -1 : 1;
     }
 
-    for (long long i = left.size() - 1; i >= 0; i--) {
+    for (long long i = static_cast<long long>(left.size()) - 1; i >= 0; i--) {
         if (left[i] != right[i]) {
             return left[i] < right[i] ? -1 : 1;
         }
